feat(autoresearch): code-split homepage AI chat components from initial load
Splits the PostHog AI chat thread and input components out of the project homepage's initial (eager) bundle:

- Extracted `ChatHeader` out of `AiFirstMaxInstance.tsx` so the homepage no longer transitively pulls in the full chat runtime (`EmbeddedRunner`, `Thread`, `Intro`) just to render the header.
- Lazy-loaded `HomepageThread` (only rendered once a chat has started) via `React.lazy` + `Suspense`.
- Extracted and lazy-loaded `HomepageAiInput` (only rendered in AI mode) via `React.lazy` + `Suspense`.

Created with Autoresearch.

## Why
The homepage always renders in idle mode first, but was eagerly bundling the full AI chat thread/input UI (Thread, QuestionInput, InputFormArea, Context, message renderers) even though those only render once a user starts an AI chat. Deferring them to a separate chunk shrinks the code needed for first paint.


Generated-By: PostHog Code
Task-Id: 5d1e0f87-d3f8-465f-969b-4242bfc1da6a
